Changed sound data types to float and int, bigger precision is not needed Refactored, moved lfo work into oscillator and gain

+++ java/src/andts/javasynth/Util.java
@@ -2,7 +2,7 @@ package andts.javasynth;
 
 public class Util
 {
-    public static byte[] trimLong(long value)
+    public static byte[] trimLong(int value)
     {
         return new byte[]{(byte) (value & 0xFF), (byte) ((value >>> 8) & 0xFF)};
     }
